feat/generic
Move shared bits to lib
Reorganize & Rename features
