+++ Cargo.toml
@@ -1,18 +1,16 @@
 [package]
 name = "command"
-version = "0.1.0"
+version = "0.2.0"
 authors = ["Lainera <fistlain@hotmail.com>"]
 edition = "2018"
 
 [features]
-default = []
-std-write = ["serde_derive"]
-rbf-write = ["ring_buffer"]
-serde_derive = ["serde", "serde_json"]
-defmt_derive = ["defmt"]
+default = ["owned"]
+owned = []
+serde_impl = ["serde", "serde_json"]
+defmt_impl = ["defmt"]
 
 [dependencies]
-ring_buffer = { git = "https://github.com/lainera/ring_buffer", optional = true }
 serde = { version = "~1.0", features = ["derive"], optional = true }
 serde_json = { version = "~1.0.63", optional = true }
-defmt = { version = "~0.2", optional = true }
+defmt = { version = "~0.3", optional = true }
